# GH-3987 introduce InternedIRI for use in Vocabularies to improve performance