--- a/core/model-api/src/main/java/org/eclipse/rdf4j/model/base/CoreDatatype.java
+++ b/core/model-api/src/main/java/org/eclipse/rdf4j/model/base/CoreDatatype.java
@@ -110,7 +110,7 @@ public interface CoreDatatype {
 		public static final String NAMESPACE = "http://www.w3.org/2001/XMLSchema#";
 
 		private static IRI iri(String localName) {
-			return new CoreDatatypeHelper.DatatypeIRI(NAMESPACE, localName);
+			return new InternedIRI(NAMESPACE, localName);
 		}
 
 		private final IRI iri;
@@ -275,7 +275,7 @@ public interface CoreDatatype {
 		public static final String NAMESPACE = "http://www.w3.org/1999/02/22-rdf-syntax-ns#";
 
 		private static IRI iri(String localName) {
-			return new CoreDatatypeHelper.DatatypeIRI(NAMESPACE, localName);
+			return new InternedIRI(NAMESPACE, localName);
 		}
 
 		private final IRI iri;
@@ -316,7 +316,7 @@ public interface CoreDatatype {
 		public static final String NAMESPACE = "http://www.opengis.net/ont/geosparql#";
 
 		private static IRI iri(String localName) {
-			return new CoreDatatypeHelper.DatatypeIRI(NAMESPACE, localName);
+			return new InternedIRI(NAMESPACE, localName);
 		}
 
 		private final IRI iri;