@hmottestad 2022-06-15
I've changed both https://github.com/eclipse/rdf4j/pull/3990 and this PR to draft for the time being. They both definitely improve performance, but that was only really obvious when combined with #3933.