@JervenBolleman 2022-06-16
Hi just a quick thought. Will this really help at scale? I have had bad experiences with intern on string similar to https://shipilev.net/jvm/anatomy-quarks/10-string-intern/