@hmottestad 2022-06-17
## Before

```
```

## After
```
Benchmark                                                                                        (param)  Mode  Cnt   Score   Error  Units
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                            moreRdfs::12180  avgt    3  11.611 ± 6.192  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                            longChain::5803  avgt    3   8.134 ± 0.454  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                                medium::544  avgt    3   0.882 ± 0.279  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                                simple::152  avgt    3   0.465 ± 0.091  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions        moreRdfs::12180  avgt    3   8.428 ± 1.013  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions        longChain::5803  avgt    3   8.121 ± 0.168  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions            medium::544  avgt    3   0.759 ± 0.233  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions            simple::152  avgt    3   0.461 ± 0.074  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema  moreRdfs::12180  avgt    3   8.738 ± 0.843  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema  longChain::5803  avgt    3   6.137 ± 0.426  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema      medium::544  avgt    3   0.514 ± 0.056  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema      simple::152  avgt    3   0.222 ± 0.033  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                      moreRdfs::12180  avgt    3   9.325 ± 3.046  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                      longChain::5803  avgt    3   6.213 ± 0.502  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                          medium::544  avgt    3   0.503 ± 0.001  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                          simple::152  avgt    3   0.220 ± 0.007  ms/op
ReasoningBenchmark.noReasoning                                                           moreRdfs::12180  avgt    3   3.979 ± 0.485  ms/op
ReasoningBenchmark.noReasoning                                                           longChain::5803  avgt    3   1.843 ± 0.124  ms/op
ReasoningBenchmark.noReasoning                                                               medium::544  avgt    3   0.297 ± 0.036  ms/op
ReasoningBenchmark.noReasoning                                                               simple::152  avgt    3   0.123 ± 0.005  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                       moreRdfs::12180  avgt    3   3.841 ± 0.864  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                       longChain::5803  avgt    3   1.838 ± 0.409  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                           medium::544  avgt    3   0.294 ± 0.036  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                           simple::152  avgt    3   0.124 ± 0.004  ms/op
```