@hmottestad 2022-06-17
## Before
```
Benchmark                                                                                        (param)  Mode  Cnt   Score   Error  Units
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                            moreRdfs::12180  avgt    5  12.385 ± 0.043  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions        moreRdfs::12180  avgt    5   9.372 ± 0.072  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema  moreRdfs::12180  avgt    5   9.447 ± 0.019  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                      moreRdfs::12180  avgt    5   9.726 ± 0.029  ms/op
ReasoningBenchmark.noReasoning                                                           moreRdfs::12180  avgt    5   3.877 ± 0.016  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                       moreRdfs::12180  avgt    5   3.858 ± 0.036  ms/op
```

## After
```
```