@hmottestad 2022-06-17
This is showing a decent 10% performance improvement in the more complex RDFS reasoning benchmark.